+++ lib/flutter_number_captcha.dart
@@ -3,34 +3,6 @@
 // To add platforms, run `flutter create -t plugin --platforms <platforms> .` under the same
 // directory. You can also find a detailed instruction on how to add platforms in the `pubspec.yaml` at https://flutter.dev/docs/development/packages-and-plugins/developing-packages#plugin-platforms.
 
-import 'package:flutter/material.dart';
+library flutter_number_captcha;
 
-import 'number_captcha_dialog.dart';
-
-class FlutterNumberCaptcha {
-  static Future<bool> show(
-    context, {
-    String titleText = 'Enter correct number',
-    String placeholderText = 'Enter Number',
-    String checkCaption = 'Check',
-    String invalidText = 'Invalid Code',
-    Color? accentColor,
-  }) async {
-    bool? result = await showDialog(
-      context: context,
-      builder: (context) {
-        return NumberCaptchaDialog(
-          titleText,
-          placeholderText,
-          checkCaption,
-          invalidText,
-          accentColor: accentColor,
-        );
-      },
-    );
-    if (result == true) {
-      return true;
-    }
-    return false;
-  }
-}
+export './src/number_captcha_dialog.dart';
