+++ pubspec.yaml
@@ -4,8 +4,7 @@ version: 0.0.8
 homepage: https://github.com/lustea00/flutter_number_captcha
 
 environment:
-  sdk: ">=2.16.1 <3.0.0"
-  flutter: ">=2.5.0"
+  sdk: '>=3.2.5 <4.0.0'
 
 dependencies:
   flutter:
@@ -14,10 +13,4 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-
-# For information on the generic Dart part of this file, see the
-# following page: https://dart.dev/tools/pub/pubspec
-
-# The following section is specific to Flutter.
-flutter:
